+++ Sources/ParseSwift/Parse.swift
@@ -126,7 +126,6 @@ public struct ParseConfiguration {
                 webhookKey: String? = nil,
                 serverURL: URL,
                 liveQueryServerURL: URL? = nil,
-                allowCustomObjectId: Bool = false,
                 allowingCustomObjectIds: Bool = false,
                 usingTransactions: Bool = false,
                 usingEqualQueryConstraint: Bool = false,
@@ -135,7 +134,6 @@ public struct ParseConfiguration {
                 requestCachePolicy: URLRequest.CachePolicy = .useProtocolCachePolicy,
                 cacheMemoryCapacity: Int = 512_000,
                 cacheDiskCapacity: Int = 10_000_000,
-                migratingFromObjcSDK: Bool = false,
                 deletingKeychainIfNeeded: Bool = false,
                 httpAdditionalHeaders: [AnyHashable: Any]? = nil,
                 maxConnectionAttempts: Int = 5,
@@ -158,11 +156,94 @@ public struct ParseConfiguration {
         self.requestCachePolicy = requestCachePolicy
         self.cacheMemoryCapacity = cacheMemoryCapacity
         self.cacheDiskCapacity = cacheDiskCapacity
-        self.isMigratingFromObjcSDK = migratingFromObjcSDK
         self.isDeletingKeychainIfNeeded = deletingKeychainIfNeeded
         self.httpAdditionalHeaders = httpAdditionalHeaders
         self.maxConnectionAttempts = maxConnectionAttempts
         ParseStorage.shared.use(keyValueStore ?? InMemoryKeyValueStore())
+    }
+
+    /**
+     Create a Parse Swift configuration.
+     - parameter applicationId: The application id for your Parse application.
+     - parameter clientKey: The client key for your Parse application.
+     - parameter masterKey: The master key for your Parse application. This key should only be
+     specified when using the SDK on a server.
+     - parameter serverURL: The server URL to connect to Parse Server.
+     - parameter liveQueryServerURL: The live query server URL to connect to Parse Server.
+     - parameter allowingCustomObjectIds: Allows objectIds to be created on the client.
+     side for each object. Must be enabled on the server to work.
+     - parameter usingTransactions: Use transactions when saving/updating multiple objects.
+     - parameter usingEqualQueryConstraint: Use the **$eq** query constraint when querying.
+     - parameter usingPostForQuery: Use **POST** instead of **GET** when making query calls.
+     Defaults to **false**.
+     - parameter keyValueStore: A key/value store that conforms to the `ParseKeyValueStore`
+     protocol. Defaults to `nil` in which one will be created an memory, but never persisted. For Linux, this
+     this is the only store available since there is no Keychain. Linux users should replace this store with an
+     encrypted one.
+     - parameter requestCachePolicy: The default caching policy for all http requests that determines
+     when to return a response from the cache. Defaults to `useProtocolCachePolicy`. See Apple's [documentation](https://developer.apple.com/documentation/foundation/url_loading_system/accessing_cached_data)
+     for more info.
+     - parameter cacheMemoryCapacity: The memory capacity of the cache, in bytes. Defaults to 512KB.
+     - parameter cacheDiskCapacity: The disk capacity of the cache, in bytes. Defaults to 10MB.
+     - parameter migratingFromObjcSDK: If your app previously used the iOS Objective-C SDK, setting this value
+     to **true** will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to **false**.
+     - parameter deletingKeychainIfNeeded: Deletes the Parse Keychain when the app is running for the first time.
+     Defaults to **false**.
+     - parameter httpAdditionalHeaders: A dictionary of additional headers to send with requests. See Apple's
+     [documentation](https://developer.apple.com/documentation/foundation/urlsessionconfiguration/1411532-httpadditionalheaders)
+     for more info.
+     - parameter maxConnectionAttempts: Maximum number of times to try to connect to Parse Server.
+     Defaults to 5.
+     - parameter authentication: A callback block that will be used to receive/accept/decline network challenges.
+     Defaults to `nil` in which the SDK will use the default OS authentication methods for challenges.
+     It should have the following argument signature: `(challenge: URLAuthenticationChallenge,
+     completionHandler: (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) -> Void`.
+     See Apple's [documentation](https://developer.apple.com/documentation/foundation/urlsessiontaskdelegate/1411595-urlsession) for more for details.
+     - warning: `usingTransactions` is experimental.
+     - warning: It is recomended to only specify `masterKey` when using the SDK on a server. Do not use this key on the client.
+     - warning: Setting `usingPostForQuery` to **true**  will require all queries to access the server instead of following the `requestCachePolicy`.
+     */
+    @available(*, deprecated, message: "Remove the migratingFromObjcSDK argument")
+    public init(applicationId: String,
+                clientKey: String? = nil,
+                masterKey: String? = nil,
+                webhookKey: String? = nil,
+                serverURL: URL,
+                liveQueryServerURL: URL? = nil,
+                allowingCustomObjectIds: Bool = false,
+                usingTransactions: Bool = false,
+                usingEqualQueryConstraint: Bool = false,
+                usingPostForQuery: Bool = false,
+                keyValueStore: ParseKeyValueStore? = nil,
+                requestCachePolicy: URLRequest.CachePolicy = .useProtocolCachePolicy,
+                cacheMemoryCapacity: Int = 512_000,
+                cacheDiskCapacity: Int = 10_000_000,
+                migratingFromObjcSDK: Bool = false,
+                deletingKeychainIfNeeded: Bool = false,
+                httpAdditionalHeaders: [AnyHashable: Any]? = nil,
+                maxConnectionAttempts: Int = 5,
+                authentication: ((URLAuthenticationChallenge,
+                                  (URLSession.AuthChallengeDisposition,
+                                   URLCredential?) -> Void) -> Void)? = nil) {
+        self.init(applicationId: applicationId,
+                  clientKey: clientKey,
+                  masterKey: masterKey,
+                  webhookKey: webhookKey,
+                  serverURL: serverURL,
+                  liveQueryServerURL: liveQueryServerURL,
+                  allowingCustomObjectIds: allowingCustomObjectIds,
+                  usingTransactions: usingTransactions,
+                  usingEqualQueryConstraint: usingEqualQueryConstraint,
+                  usingPostForQuery: usingPostForQuery,
+                  keyValueStore: keyValueStore,
+                  requestCachePolicy: requestCachePolicy,
+                  cacheMemoryCapacity: cacheMemoryCapacity,
+                  cacheDiskCapacity: cacheDiskCapacity,
+                  deletingKeychainIfNeeded: deletingKeychainIfNeeded,
+                  httpAdditionalHeaders: httpAdditionalHeaders,
+                  maxConnectionAttempts: maxConnectionAttempts,
+                  authentication: authentication)
+        self.isMigratingFromObjcSDK = migratingFromObjcSDK
     }
 }
 
@@ -193,7 +274,7 @@ public struct ParseSwift {
             // All migrations from previous versions to current should occur here:
             #if !os(Linux) && !os(Android) && !os(Windows)
             if previousSDKVersion < oneNineEightSDKVersion {
-                // Old macOS Keychain cannot be used because it's global to all apps.
+                // Old macOS Keychain cannot be used because it is global to all apps.
                 _ = KeychainStore.old
                 KeychainStore.shared.copy(keychain: KeychainStore.old)
                 // Need to delete the old Keychain because a new one is created with bundleId.
@@ -274,8 +355,6 @@ public struct ParseSwift {
      for more info.
      - parameter cacheMemoryCapacity: The memory capacity of the cache, in bytes. Defaults to 512KB.
      - parameter cacheDiskCapacity: The disk capacity of the cache, in bytes. Defaults to 10MB.
-     - parameter migratingFromObjcSDK: If your app previously used the iOS Objective-C SDK, setting this value
-     to **true** will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to **false**.
      - parameter deletingKeychainIfNeeded: Deletes the Parse Keychain when the app is running for the first time.
      Defaults to **false**.
      - parameter httpAdditionalHeaders: A dictionary of additional headers to send with requests. See Apple's
@@ -304,6 +383,88 @@ public struct ParseSwift {
         requestCachePolicy: URLRequest.CachePolicy = .useProtocolCachePolicy,
         cacheMemoryCapacity: Int = 512_000,
         cacheDiskCapacity: Int = 10_000_000,
+        deletingKeychainIfNeeded: Bool = false,
+        httpAdditionalHeaders: [AnyHashable: Any]? = nil,
+        maxConnectionAttempts: Int = 5,
+        authentication: ((URLAuthenticationChallenge,
+                          (URLSession.AuthChallengeDisposition,
+                           URLCredential?) -> Void) -> Void)? = nil
+    ) {
+        let configuration = ParseConfiguration(applicationId: applicationId,
+                                               clientKey: clientKey,
+                                               masterKey: masterKey,
+                                               serverURL: serverURL,
+                                               liveQueryServerURL: liveQueryServerURL,
+                                               allowingCustomObjectIds: allowingCustomObjectIds,
+                                               usingTransactions: usingTransactions,
+                                               usingEqualQueryConstraint: usingEqualQueryConstraint,
+                                               usingPostForQuery: usingPostForQuery,
+                                               keyValueStore: keyValueStore,
+                                               requestCachePolicy: requestCachePolicy,
+                                               cacheMemoryCapacity: cacheMemoryCapacity,
+                                               cacheDiskCapacity: cacheDiskCapacity,
+                                               deletingKeychainIfNeeded: deletingKeychainIfNeeded,
+                                               httpAdditionalHeaders: httpAdditionalHeaders,
+                                               maxConnectionAttempts: maxConnectionAttempts,
+                                               authentication: authentication)
+        initialize(configuration: configuration)
+    }
+
+    /**
+     Configure the Parse Swift client. This should only be used when starting your app. Typically in the
+     `application(... didFinishLaunchingWithOptions launchOptions...)`.
+     - parameter applicationId: The application id for your Parse application.
+     - parameter clientKey: The client key for your Parse application.
+     - parameter masterKey: The master key for your Parse application. This key should only be
+     specified when using the SDK on a server.
+     - parameter serverURL: The server URL to connect to Parse Server.
+     - parameter liveQueryServerURL: The live query server URL to connect to Parse Server.
+     - parameter allowingCustomObjectIds: Allows objectIds to be created on the client.
+     side for each object. Must be enabled on the server to work.
+     - parameter usingTransactions: Use transactions when saving/updating multiple objects.
+     - parameter usingEqualQueryConstraint: Use the **$eq** query constraint when querying.
+     - parameter usingPostForQuery: Use **POST** instead of **GET** when making query calls.
+     Defaults to **false**.
+     - parameter keyValueStore: A key/value store that conforms to the `ParseKeyValueStore`
+     protocol. Defaults to `nil` in which one will be created an memory, but never persisted. For Linux, this
+     this is the only store available since there is no Keychain. Linux users should replace this store with an
+     encrypted one.
+     - parameter requestCachePolicy: The default caching policy for all http requests that determines
+     when to return a response from the cache. Defaults to `useProtocolCachePolicy`. See Apple's [documentation](https://developer.apple.com/documentation/foundation/url_loading_system/accessing_cached_data)
+     for more info.
+     - parameter cacheMemoryCapacity: The memory capacity of the cache, in bytes. Defaults to 512KB.
+     - parameter cacheDiskCapacity: The disk capacity of the cache, in bytes. Defaults to 10MB.
+     - parameter migratingFromObjcSDK: If your app previously used the iOS Objective-C SDK, setting this value
+     to **true** will attempt to migrate relevant data stored in the Keychain to ParseSwift. Defaults to **false**.
+     - parameter deletingKeychainIfNeeded: Deletes the Parse Keychain when the app is running for the first time.
+     Defaults to **false**.
+     - parameter httpAdditionalHeaders: A dictionary of additional headers to send with requests. See Apple's
+     [documentation](https://developer.apple.com/documentation/foundation/urlsessionconfiguration/1411532-httpadditionalheaders)
+     for more info.
+     - parameter authentication: A callback block that will be used to receive/accept/decline network challenges.
+     Defaults to `nil` in which the SDK will use the default OS authentication methods for challenges.
+     It should have the following argument signature: `(challenge: URLAuthenticationChallenge,
+     completionHandler: (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) -> Void`.
+     See Apple's [documentation](https://developer.apple.com/documentation/foundation/urlsessiontaskdelegate/1411595-urlsession) for more for details.
+     - warning: `usingTransactions` is experimental.
+     - warning: It is recomended to only specify `masterKey` when using the SDK on a server. Do not use this key on the client.
+     - warning: Setting `usingPostForQuery` to **true**  will require all queries to access the server instead of following the `requestCachePolicy`.
+     */
+    @available(*, deprecated, message: "Remove the migratingFromObjcSDK argument")
+    static public func initialize(
+        applicationId: String,
+        clientKey: String? = nil,
+        masterKey: String? = nil,
+        serverURL: URL,
+        liveQueryServerURL: URL? = nil,
+        allowingCustomObjectIds: Bool = false,
+        usingTransactions: Bool = false,
+        usingEqualQueryConstraint: Bool = false,
+        usingPostForQuery: Bool = false,
+        keyValueStore: ParseKeyValueStore? = nil,
+        requestCachePolicy: URLRequest.CachePolicy = .useProtocolCachePolicy,
+        cacheMemoryCapacity: Int = 512_000,
+        cacheDiskCapacity: Int = 10_000_000,
         migratingFromObjcSDK: Bool = false,
         deletingKeychainIfNeeded: Bool = false,
         httpAdditionalHeaders: [AnyHashable: Any]? = nil,
@@ -312,24 +473,25 @@ public struct ParseSwift {
                           (URLSession.AuthChallengeDisposition,
                            URLCredential?) -> Void) -> Void)? = nil
     ) {
-        initialize(configuration: .init(applicationId: applicationId,
-                                        clientKey: clientKey,
-                                        masterKey: masterKey,
-                                        serverURL: serverURL,
-                                        liveQueryServerURL: liveQueryServerURL,
-                                        allowingCustomObjectIds: allowingCustomObjectIds,
-                                        usingTransactions: usingTransactions,
-                                        usingEqualQueryConstraint: usingEqualQueryConstraint,
-                                        usingPostForQuery: usingPostForQuery,
-                                        keyValueStore: keyValueStore,
-                                        requestCachePolicy: requestCachePolicy,
-                                        cacheMemoryCapacity: cacheMemoryCapacity,
-                                        cacheDiskCapacity: cacheDiskCapacity,
-                                        migratingFromObjcSDK: migratingFromObjcSDK,
-                                        deletingKeychainIfNeeded: deletingKeychainIfNeeded,
-                                        httpAdditionalHeaders: httpAdditionalHeaders,
-                                        maxConnectionAttempts: maxConnectionAttempts,
-                                        authentication: authentication))
+        var configuration = ParseConfiguration(applicationId: applicationId,
+                                               clientKey: clientKey,
+                                               masterKey: masterKey,
+                                               serverURL: serverURL,
+                                               liveQueryServerURL: liveQueryServerURL,
+                                               allowingCustomObjectIds: allowingCustomObjectIds,
+                                               usingTransactions: usingTransactions,
+                                               usingEqualQueryConstraint: usingEqualQueryConstraint,
+                                               usingPostForQuery: usingPostForQuery,
+                                               keyValueStore: keyValueStore,
+                                               requestCachePolicy: requestCachePolicy,
+                                               cacheMemoryCapacity: cacheMemoryCapacity,
+                                               cacheDiskCapacity: cacheDiskCapacity,
+                                               deletingKeychainIfNeeded: deletingKeychainIfNeeded,
+                                               httpAdditionalHeaders: httpAdditionalHeaders,
+                                               maxConnectionAttempts: maxConnectionAttempts,
+                                               authentication: authentication)
+        configuration.isMigratingFromObjcSDK = migratingFromObjcSDK
+        initialize(configuration: configuration)
     }
 
     internal static func initialize(applicationId: String,
@@ -366,11 +528,11 @@ public struct ParseSwift {
                                                requestCachePolicy: requestCachePolicy,
                                                cacheMemoryCapacity: cacheMemoryCapacity,
                                                cacheDiskCapacity: cacheDiskCapacity,
-                                               migratingFromObjcSDK: migratingFromObjcSDK,
                                                deletingKeychainIfNeeded: deletingKeychainIfNeeded,
                                                httpAdditionalHeaders: httpAdditionalHeaders,
                                                maxConnectionAttempts: maxConnectionAttempts,
                                                authentication: authentication)
+        configuration.isMigratingFromObjcSDK = migratingFromObjcSDK
         configuration.isTestingSDK = testing
         initialize(configuration: configuration)
     }